requirements.txt
flask
gunicorn
yfinance
textblob
google.generativeai